Add HomeController, Exceptions for Customer/City/Warehouse

+++ DeliverIT.API/Controllers/CustomerController.cs
@@ -1,9 +1,7 @@
 ﻿using DeliverIT.Services.Contracts;
 using DeliverIT.Services.DTOs;
 using DeliverIT.Services.Helpers;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -15,11 +13,9 @@ namespace DeliverIT.API.Controllers
     public class CustomerController : ControllerBase
     {
         private readonly ICustomerService _cs;
-        private readonly IAuthenticationService _auth;
-        public CustomerController(ICustomerService cs, IAuthenticationService auth)
+        public CustomerController(ICustomerService cs)
         {
             this._cs = cs;
-            this._auth = auth;
         }
 
         // must be public - non authorisation needed
@@ -28,36 +24,6 @@ namespace DeliverIT.API.Controllers
         public async Task<ActionResult> CustomerCount()
         {
             return Ok(await _cs.UserCountAsync());
-        }
-
-        [HttpGet("login")]
-        public async Task<ActionResult> Login([FromHeader] string credentials)
-        {
-            var isUserExisting = this._auth.FindUser(credentials);
-            if (isUserExisting)
-            {
-                var userId = await this._auth.GetUserID(credentials);
-                var options = new CookieOptions
-                {
-                    Expires = DateTime.Now.AddMinutes(5)
-                };
-                this.Response.Cookies.Append(Constants.KEY_USER_ID, userId, options);
-                return this.Ok(Constants.LOGGED);
-            }
-
-            var isEmployeeExisting = this._auth.FindEmployee(credentials);
-            if (isEmployeeExisting)
-            {
-                var employeeId = await this._auth.GetEmployeeID(credentials);
-                var options = new CookieOptions
-                {
-                    Expires = DateTime.Now.AddMinutes(5)
-                };
-                this.Response.Cookies.Append(Constants.KEY_EMPLOYEE_ID, employeeId, options);
-                return this.Ok(Constants.LOGGED);
-            }
-
-            return this.Unauthorized();
         }
 
         [HttpGet("all")]
@@ -145,15 +111,6 @@ namespace DeliverIT.API.Controllers
             return this.Ok(customers);
         }
 
-        //must be public
-        [HttpPost("register")]
-        [ProducesResponseType(201)]
-        [ProducesResponseType(400)]
-        public async Task<ActionResult<CustomerDTO>> CreateCustomerAsync(CustomerDTO obj)
-        {
-            return this.Created("Get", await this._cs.PostAsync(obj));
-        }
-
         [HttpDelete("{id}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(404)]
@@ -172,7 +129,7 @@ namespace DeliverIT.API.Controllers
             }
             else if (this.Request.Cookies.ContainsKey(Constants.KEY_USER_ID))
             {
-                if (id == int.Parse(Request.Cookies[Constants.KEY_USER_ID]))
+                if (Request.Cookies[Constants.KEY_USER_ID] != null && id == int.Parse(Request.Cookies[Constants.KEY_USER_ID]))
                 {
                     return this.Ok(await this._cs.DeleteAsync(id));
                 }
